pyproject.toml
[tool.poetry]
name = "langchain-rag-demo"
version = "0.1.0"
description = "A RAG (Retrieval-Augmented Generation) demo project using LangChain, Ollama and Qdrant"
authors = ["Polin"]
packages = [
    {include = "main.py"},
    {include = "*.py"}
]
include = ["*.py", "*.json", "*.md"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11,<3.12"
python-dotenv = "^1.0.0"
qdrant-client = "^1.12.0"
langchain-community = "^0.2.19"
langchain = "^0.2.17"
langchain-qdrant = "^0.1.4"
langchain-core = "^0.2.43"
langchain-text-splitters = "^0.2.1"
pypdf = "^4.3.0"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
python-multipart = "^0.0.9"
pydantic = "^2.7.4"
jinja2 = "^3.1.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
